refactored cucumber model namespace
