+++ org.agileware.natural.cucumber.tests/src/org/agileware/natural/cucumber/tests/CucumberExamplesTest.xtend
@@ -1,6 +1,6 @@
 package org.agileware.natural.cucumber.tests
 
-import org.agileware.natural.cucumber.cucumber.Feature
+import org.agileware.natural.cucumber.model.Feature
 import org.agileware.natural.testing.AbstractExamplesTest
 import org.eclipse.xtext.testing.InjectWith
 import org.eclipse.xtext.testing.XtextRunner
