+++ org.agileware.natural.cucumber.ui/src/org/agileware/natural/cucumber/ui/labeling/CucumberLabelProvider.java
@@ -3,15 +3,15 @@
  */
 package org.agileware.natural.cucumber.ui.labeling;
 
-import org.agileware.natural.cucumber.cucumber.Background;
-import org.agileware.natural.cucumber.cucumber.DocString;
-import org.agileware.natural.cucumber.cucumber.Example;
-import org.agileware.natural.cucumber.cucumber.Feature;
-import org.agileware.natural.cucumber.cucumber.Scenario;
-import org.agileware.natural.cucumber.cucumber.ScenarioOutline;
-import org.agileware.natural.cucumber.cucumber.Step;
-import org.agileware.natural.cucumber.cucumber.Table;
-import org.agileware.natural.cucumber.cucumber.Tag;
+import org.agileware.natural.cucumber.model.Background;
+import org.agileware.natural.cucumber.model.DocString;
+import org.agileware.natural.cucumber.model.Example;
+import org.agileware.natural.cucumber.model.Feature;
+import org.agileware.natural.cucumber.model.Scenario;
+import org.agileware.natural.cucumber.model.ScenarioOutline;
+import org.agileware.natural.cucumber.model.Step;
+import org.agileware.natural.cucumber.model.Table;
+import org.agileware.natural.cucumber.model.Tag;
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.edit.ui.provider.AdapterFactoryLabelProvider;
 import org.eclipse.xtext.ui.label.DefaultEObjectLabelProvider;
