+++ org.agileware.natural.cucumber/META-INF/MANIFEST.MF
@@ -19,20 +19,20 @@ Require-Bundle: org.agileware.natural.common;visibility:=reexport,
  org.antlr.runtime;bundle-version="[3.2.0,3.2.1)"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.agileware.natural.cucumber;uses:="com.google.inject,org.eclipse.xtext.service,org.eclipse.xtext",
- org.agileware.natural.cucumber.cucumber;uses:="org.eclipse.emf.ecore,org.eclipse.emf.common.util",
- org.agileware.natural.cucumber.cucumber.impl;
+ org.agileware.natural.cucumber.model;uses:="org.eclipse.emf.ecore,org.eclipse.emf.common.util",
+ org.agileware.natural.cucumber.model.impl;
   uses:="org.eclipse.emf.ecore,
    org.eclipse.emf.ecore.impl,
    org.eclipse.emf.common.util,
    org.eclipse.emf.common.notify,
-   org.agileware.natural.cucumber.cucumber",
- org.agileware.natural.cucumber.cucumber.util;
+   org.agileware.natural.cucumber.model",
+ org.agileware.natural.cucumber.model.util;
   uses:="org.eclipse.emf.ecore,
    org.eclipse.emf.common.notify,
    org.eclipse.emf.common.notify.impl,
    org.eclipse.emf.ecore.util,
-   org.agileware.natural.cucumber.cucumber",
- org.agileware.natural.cucumber.formatting2;uses:="org.eclipse.xtext.formatting2,org.agileware.natural.cucumber.cucumber",
+   org.agileware.natural.cucumber.model",
+ org.agileware.natural.cucumber.formatting2;uses:="org.eclipse.xtext.formatting2,org.agileware.natural.cucumber.model",
  org.agileware.natural.cucumber.generator;uses:="org.eclipse.xtext.generator,org.eclipse.emf.ecore.resource",
  org.agileware.natural.cucumber.parser.antlr;uses:="org.eclipse.xtext.parser.antlr,org.agileware.natural.cucumber.parser.antlr.internal,org.agileware.natural.cucumber.services",
  org.agileware.natural.cucumber.parser.antlr.internal;
@@ -41,7 +41,10 @@ Export-Package: org.agileware.natural.cucumber;uses:="com.google.inject,org.ecli
    org.antlr.runtime,
    org.agileware.natural.cucumber.services",
  org.agileware.natural.cucumber.scoping;uses:="org.eclipse.xtext.scoping.impl",
- org.agileware.natural.cucumber.serializer;uses:="org.agileware.natural.cucumber.cucumber",
+ org.agileware.natural.cucumber.serializer;uses:="org.agileware.natural.cucumber.model",
  org.agileware.natural.cucumber.services;uses:="org.eclipse.xtext.service,org.eclipse.xtext",
- org.agileware.natural.cucumber.validation;uses:="org.eclipse.xtext.util,org.eclipse.xtext.validation"
+ org.agileware.natural.cucumber.validation;uses:="org.eclipse.xtext.util,org.eclipse.xtext.validation",
+ org.agileware.natural.cucumber.model.impl,
+ org.agileware.natural.cucumber.model.util,
+ org.agileware.natural.cucumber.model
 Import-Package: org.apache.log4j
