+++ org.agileware.natural.cucumber/src/org/agileware/natural/cucumber/Cucumber.xtext
@@ -1,7 +1,7 @@
 grammar org.agileware.natural.cucumber.Cucumber hidden(WS, SL_COMMENT)
 
 import "http://www.eclipse.org/emf/2002/Ecore" as ecore
-generate cucumber "http://www.agileware.org/natural/cucumber"
+generate model "http://www.agileware.org/natural/cucumber"
 
 Feature:
 	tags+=Tag*
@@ -168,7 +168,7 @@ terminal WORD:
 terminal WS: (' ' | '\t');
 
 terminal EOL: NL;
-terminal fragment NL: ('\r'? '\n'?);
+terminal fragment NL: ('\r'? '\n');
 
 terminal ANY_OTHER: .;
