+++ org.agileware.natural.cucumber/src/org/agileware/natural/cucumber/serializer/CucumberSerializer.xtend
@@ -1,18 +1,18 @@
 package org.agileware.natural.cucumber.serializer
 
-import org.agileware.natural.cucumber.cucumber.Background
-import org.agileware.natural.cucumber.cucumber.DocString
-import org.agileware.natural.cucumber.cucumber.Example
-import org.agileware.natural.cucumber.cucumber.Feature
-import org.agileware.natural.cucumber.cucumber.Scenario
-import org.agileware.natural.cucumber.cucumber.ScenarioOutline
-import org.agileware.natural.cucumber.cucumber.Step
-import org.agileware.natural.cucumber.cucumber.Table
-import org.agileware.natural.cucumber.cucumber.TableCol
-import org.agileware.natural.cucumber.cucumber.TableRow
-import org.agileware.natural.cucumber.cucumber.Tag
+import org.agileware.natural.cucumber.model.Background
+import org.agileware.natural.cucumber.model.DocString
+import org.agileware.natural.cucumber.model.Example
+import org.agileware.natural.cucumber.model.Feature
+import org.agileware.natural.cucumber.model.Scenario
+import org.agileware.natural.cucumber.model.ScenarioOutline
+import org.agileware.natural.cucumber.model.Step
+import org.agileware.natural.cucumber.model.Table
+import org.agileware.natural.cucumber.model.TableCol
+import org.agileware.natural.cucumber.model.TableRow
+import org.agileware.natural.cucumber.model.Tag
 import org.eclipse.xtext.nodemodel.util.NodeModelUtils
-import org.agileware.natural.cucumber.cucumber.Text
+import org.agileware.natural.cucumber.model.Text
 
 class CucumberSerializer {
